Remove the native multiplatform extension (#28)
Native multiplatform extension was responsive for
easing native targets and source set setup.
Previously, relying on KGP facilities to set up native
targets was a tedious process.
With the latest KGP versions that is no longer true.

Moreover, the API this extension was relying on is
deprecated and will be removed in Kotlin 2.2.
The extension could be rewritten to use a supported API,
but it does make a lot of sense, as almost all projects
relying on infra plugin stopped using target-related
machinery some time ago.