Don't apply signing plugin if signing is not configured
Currently, we're always applying the signing plugin (when configuring signing),
but then, if the PGP key is unknown,
we abstain from further configuration.

Such an approach works fine, until it doesn't:
Gradle "plugin publish" plugin setup signing (and creates sign-task) if
the signing plugin was applied to a project,
and then the task fails due to missing credentials.

This change postpones signing plugin application until we're sure that credentials were provided.